feat: implement snake gameplay architecture, linked list body, obstacle system, and data-driven level flow

1. Implemented PlayerService and SnakeController with lifecycle management, global integration, and controlled update/render execution during GAMEPLAY.
2. Designed SnakeController state system with ALIVE and DEAD flow, restart and respawn handling, input stability, and delayed time-based movement.
3. Added direction and position framework including grid-based movement, screen wrapping, and precision fixes in grid and cell calculations.
4. Implemented BodyPart class with rendering, lifecycle, rotation, movement helpers, and grid-to-screen positioning.
5. Built dynamic snake body using a single linked list, including node structure, head and body spawning, direction propagation, and movement update.
6. Implemented safe memory management with node removal, cleanup, and snake reset functionality.
7. Added prediction-based body collision detection and movement safety to improve gameplay stability.
8. Created modular obstacle system with ElementData, Obstacle class, and lifecycle management.
9. Implemented ElementService with dynamic obstacle spawning, update, rendering, and ServiceLocator integration.
10. Designed scalable data-driven level architecture supporting element configuration and automatic spawning during level loading.
11. Integrated level pipeline with element system and configured level two obstacle layout.
12. Established a modular and extensible gameplay foundation for future systems such as food, traps, enemies, and power-ups.

+++ Linked-List-Snake/include/Level/LevelView.h
@@ -37,7 +37,7 @@ namespace Level
 		void update();
 		void render();
 
-		int getGridWidth();
-		int getGridHeight();
+		float getGridWidth();
+		float getGridHeight();
 	};
 }
